+++ src/main/java/Temp.java
@@ -9,9 +9,9 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.*;
 import java.net.URL;
+import java.sql.*;
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Properties;
 
 public class Temp {
 
@@ -32,81 +32,16 @@ public class Temp {
         return sp.toString();
     }
 
+
     public static void main(String[] args) {
-        exampleTest();
+        Statement st = getStatementMSSQL();
+//        exampleTest();
     }
 
     public static void exampleTest() {
 //        Map<String, Map<String, String>> gMap = getNotification("src/main/resources/example.xml");
-        Map<String, Map<String, String>> gMap = getNotification(Option.getURL((byte) 2, "20211102", "20211103", null, null));
 //        Map<String, Map<String, String>> gMap = getNotification("https://torgi.gov.ru/opendata/notification/55829150");
-        addDetailInfo(gMap);
-        printNumber(gMap, "173");
-    }
 
-    public static void printAll(Map<String, Map<String, String>> gMap) {
-        //Print global map
-        for (Map.Entry<String, Map<String, String>> gm : gMap.entrySet()) {
-            for (Map.Entry<String, String> m : gm.getValue().entrySet()) {
-                System.out.println("Num " + gm.getKey() + " -- Key: " + m.getKey() + ", value: " + m.getValue());
-            }
-        }
-    }
-
-    public static void printNumber(Map<String, Map<String, String>> gMap, String idNotificaton) {
-        //Print specify idNotificaton
-        for (Map.Entry<String, Map<String, String>> gm : gMap.entrySet()) {
-            if (gm.getKey().startsWith(idNotificaton) || gm.getKey().startsWith(idNotificaton + "-")) {
-                for (Map.Entry<String, String> m : gm.getValue().entrySet()) {
-                    System.out.println("Num " + gm.getKey() + " -- Key: " + m.getKey() + ", value: " + m.getValue());
-                }
-            }
-        }
-    }
-
-    public static Map<String, Map<String, String>> getNotification(String url) {
-
-        //Global map of notifications
-        Map<String, Map<String, String>> gMap = new HashMap<>();
-
-        Document doc = Option.getDoc(url);
-        if (doc != null) {
-            //Get main list of notification
-            NodeList mainNodeList = doc.getElementsByTagName("notification");
-            for (int i = 0; i < mainNodeList.getLength(); i++) {
-                Node node = mainNodeList.item(i);
-                //Get all tag (key, values) in node i
-                Map<String, String> map = getTag(node);
-                gMap.put(String.valueOf(i), map);
-            }
-        }
-        return gMap;
-    }
-
-    public static void addDetailInfo(Map<String, Map<String, String>> gMap) {
-
-        Map<String, String> temp = new HashMap<>();
-        for (Map.Entry<String, Map<String, String>> gm : gMap.entrySet()) {
-            String idNotificaton = gm.getKey();
-            //Check availability tag 'odDetailedHref'
-            if (gm.getValue().containsKey("odDetailedHref")) {
-                //Get doc odDetailedHref
-                String url = gm.getValue().get("odDetailedHref").replace("http://", "https://");
-                temp.put(idNotificaton, url);
-            }
-        }
-
-        for (Map.Entry<String, String> e : temp.entrySet()) {
-            Document detailDoc = Option.getDoc(e.getValue());
-            if (detailDoc != null) {
-                NodeList detailNodeList = detailDoc.getElementsByTagName("notification");
-                for (int i = 0; i < detailNodeList.getLength(); i++) {
-                    Node detailNode = detailNodeList.item(i);
-                    Map<String, String> detailMap = getTag(detailNode);
-                    gMap.put(e.getKey() + "-" + i, detailMap);
-                }
-            }
-        }
     }
 
     public static void getTagDetail(Node node) {
@@ -114,7 +49,7 @@ public class Temp {
     }
 
     public static void getTagDetail(Node node, String tag) {
-        if ((node.getNodeType() == Node.TEXT_NODE) && (hasText(node.getTextContent()))) {
+        if ((node.getNodeType() == Node.TEXT_NODE) && (LoadData.hasText(node.getTextContent()))) {
 //            System.out.println("Node '" + node.getNodeName() + "' (type " + node.getNodeType() + ") has " +
 //                    node.getChildNodes().getLength() + " child");
 //        if (node.getNodeType() == Node.ELEMENT_NODE) {
@@ -131,10 +66,10 @@ public class Temp {
                     + node.getNodeType() + ") (child: "
                     + node.getChildNodes().getLength() + ") (nodeName: "
                     + node.getNodeName() + ") "
-                    + clearText(node.getTextContent()));
+                    + LoadData.clearText(node.getTextContent()));
             System.out.println("--------------------------------------------------");
         } else {
-            ArrayList<Node> nodeList = clearNodeList(node.getChildNodes());
+            ArrayList<Node> nodeList = LoadData.clearNodeList(node.getChildNodes());
             System.out.println("Size of Array: " + nodeList.size());
             for (int i = 0; i < nodeList.size(); i++) {
                 System.out.println("Go from node '" + node.getParentNode().getNodeName() + "' to node №" + (i + 1)
@@ -143,43 +78,6 @@ public class Temp {
                 getTagDetail(nodeList.get(i), tag + "/" + nodeList.get(i).getParentNode().getNodeName());
             }
         }
-    }
-
-    public static Map<String, String> getTag(Node node) {
-        return getTag(node, "", new HashMap<>());
-    }
-
-    public static Map<String, String> getTag(Node node, String tag, Map<String, String> map) {
-        if ((node.getNodeType() == Node.TEXT_NODE) && (hasText(node.getTextContent()))) {
-            //System.out.println(tag.substring(14) + ": " + clearText(node.getTextContent()));
-            map.put(tag.substring(14), clearText(node.getTextContent()));
-        } else {
-            ArrayList<Node> nodeList = clearNodeList(node.getChildNodes());
-            for (Node value : nodeList) {
-                getTag(value, tag + "/" + value.getParentNode().getNodeName(), map);
-            }
-        }
-        return map;
-    }
-
-    public static boolean hasText(String str) {
-        return str.replaceAll("\n", "").replaceAll("\t", "")
-                .replaceAll(" ", "").length() != 0;
-    }
-
-    public static String clearText(String str) {
-        str = str.replaceAll("\n", "").replaceAll("\t", "");
-        while (str.contains("  ")) str = str.replaceAll(" {2}", " ");
-        return str;
-    }
-
-    public static ArrayList<Node> clearNodeList(NodeList nodeList) {
-        ArrayList<Node> newNL = new ArrayList<>();
-        for (int i = 0; i < nodeList.getLength(); i++) {
-            Node node = nodeList.item(i);
-            if ((node.getNodeType() != Node.TEXT_NODE) || (hasText(node.getTextContent()))) newNL.add(node);
-        }
-        return newNL;
     }
 
     public static void listOfTags(Document doc, String tag) {
@@ -245,6 +143,41 @@ public class Temp {
         Document doc = db.parse(url.openStream());
         NodeList nodes = doc.getElementsByTagName("*");
         System.out.println(nodes.getLength() + " nodes found");
+
+    }
+
+    public static Statement getStatementMSSQL() {
+
+        try {
+            FileReader reader = new FileReader("src/main/resources/META-INF/db.properties");
+            Properties p = new Properties();
+            p.load(reader);
+            String url = p.getProperty("urlMSSQL");
+            String user = p.getProperty("user");
+            String password = p.getProperty("password");
+            String driver = p.getProperty("driverMSSQL");
+
+            Class.forName(driver);
+
+            Connection con = DriverManager.getConnection(url, user, password);
+
+            // Create and execute a SELECT SQL statement
+            Statement statement = con.createStatement();
+            ResultSet resultSet;
+            String selectSql = "SELECT * from notifications";
+            resultSet = statement.executeQuery(selectSql);
+
+            // Print results from select statement
+            while (resultSet.next()) {
+                System.out.println(resultSet.getInt(1));
+            }
+
+            return statement;
+
+        } catch (IOException | ClassNotFoundException | SQLException e) {
+            e.printStackTrace();
+            return null;
+        }
 
     }
 }
