+++ src/main/resources/META-INF/db.properties
@@ -1,4 +1,6 @@
-driver = com.mysql.cj.jdbc.Driver
-url = jdbc:mysql://localhost:3306/torgi
+driverMySQL = com.mysql.cj.jdbc.Driver
+driverMSSQL = com.microsoft.sqlserver.jdbc.SQLServerDriver
+urlMySQL = jdbc:mysql://localhost:3306/torgi
+urlMSSQL = jdbc:sqlserver://DESKTOP-DIEJL0L\\SQLEXPRESS;databaseName=TORGI
 user = root
 password = admin
